naughty: display the client-supplied localized action description in UI
According to the Desktop Notification specification document [1] the
clients supply actions available along with a notification in a form of
a list of pairs where first element is an identifier of an action and
the second is a localized message that will be displayed to the user.

Up to now the naughty code directly used the action identifier text as a
part of the notification layout exposed to the user. This commit makes
use of a localized action description for that purpose.

Please note that this commit introduces some API breakage (the layout of
args.actions table in naughty.notify was changed to be able to pass the
action description).

1) https://developer.gnome.org/notification-spec/