Move the XDG icon theme implementation from menubar to beautiful
In order to merge and streamline all of the present icon theme
facilities in the current codebase we move the latest (0.12) XDG Icon
Theme specification implementation relevant modules out of the menubar
directory to make it logically available to the rest of the world.

The move of icon theming modules to beautiful directory is justified by
an idea that beautiful is supposed to provide some theming support in
the first place.