beautiful: icon_theme: try fallback lookup without any extensions
As the inline comment states it could be useful for some legacy
applications which specify an icon like Icon=name.png in their desktop
files.

This additional check is not supposed to introduce much of the
additional I/O overhead since it is done at a very last resort and the
number of such applications is miniscule anyway.