beautiful: icon_theme: try to use GTK+ faster implementation first
This commit makes the module use some well known external optimized and
XDG-compliant icon theme support engine if it is available. If the
library is not available (e.g. not installed) or fails a fallback to the
native Lua implementation is provided.

The fallback to the original implementation is crucial a) to make sure a
hard dependency on GTK+ won't get introduced and b) to still handle all
of the corner cases since GTK+ doesn't bother with doing it.