Revert the whole GTK+ thing for now
Well, it works and works smoothly but it breaks neat headless test suite
just the way as it is, because lgi's Gtk module feels much obliged to
pass away on an internal assertion immediately when X11 display is not
present. So the bundled unit test for icon_theme needs to be converted
into real integration test first. That wasn't part of the plan; not to
mention the original idea of mixing two providers into one wasn't
supposed to fly anyway.